+++ src/main/datatypes/MQ2MacroQuestType.cpp
@@ -24,6 +24,7 @@ enum class MacroQuestMembers
 	MQ2DataError,
 	BuildDate,
 	Build,
+	BuildName,
 	Path,
 	Version,
 	InternalName,
@@ -38,6 +39,7 @@ MQ2MacroQuestType::MQ2MacroQuestType() : MQ2Type("macroquest")
 	ScopedTypeMember(MacroQuestMembers, MQ2DataError);
 	ScopedTypeMember(MacroQuestMembers, BuildDate);
 	ScopedTypeMember(MacroQuestMembers, Build);
+	ScopedTypeMember(MacroQuestMembers, BuildName);
 	ScopedTypeMember(MacroQuestMembers, Path);
 	ScopedTypeMember(MacroQuestMembers, Version);
 	ScopedTypeMember(MacroQuestMembers, InternalName);
@@ -98,6 +100,12 @@ bool MQ2MacroQuestType::GetMember(MQVarPtr VarPtr, const char* Member, char* Ind
 	case MacroQuestMembers::Build:
 		Dest.DWord = gBuild;
 		Dest.Type = pIntType;
+		return true;
+
+	case MacroQuestMembers::BuildName:
+		strcpy_s(DataTypeTemp, GetBuildTargetName(static_cast<BuildTarget>(gBuild)));
+		Dest.Ptr = &DataTypeTemp[0];
+		Dest.Type = pStringType;
 		return true;
 
 	case MacroQuestMembers::Path: